suppression des elements inutiles dans l'appli

+++ Project/adminBlazor/adminBlazor/Pages/List.razor.cs
@@ -38,25 +38,6 @@ namespace adminBlazor.Pages
         public IWebHostEnvironment WebHostEnvironment { get; set; }
 
 
-        /*        protected override async Task OnAfterRenderAsync(bool firstRender)
-                {
-                    // Do not treat this action if is not the first render
-                    if (!firstRender)
-                    {
-                        return;
-                    }
-
-                    var currentData = await LocalStorage.GetItemAsync<User[]>("data");
-
-                    // Check if data exist in the local storage
-                    if (currentData == null)
-                    {
-                        // this code add in the local storage the fake data (we load the data sync for initialize the data before load the OnReadData method)
-                        var originalData = Http.GetFromJsonAsync<User[]>($"{NavigationManager.BaseUri}user.json").Result;
-                        await LocalStorage.SetItemAsync("data", originalData);
-                    }
-                }
-        */
 
         private async Task OnReadData(DataGridReadDataEventArgs<User> e)
         {
@@ -64,10 +45,6 @@ namespace adminBlazor.Pages
             {
                 return;
             }
-
-            // When you use a real API, we use this follow code
-            //var response = await Http.GetJsonAsync<Data[]>( $"http://my-api/api/data?page={e.Page}&pageSize={e.PageSize}" );
-            //var response = (await LocalStorage.GetItemAsync<User[]>("data")).Skip((e.Page - 1) * e.PageSize).Take(e.PageSize).ToList();
 
             if (!e.CancellationToken.IsCancellationRequested)
             {
@@ -84,13 +61,15 @@ namespace adminBlazor.Pages
             var modal = Modal.Show<UserDeleteConfirmation>("Delete Confirmation", parameters);
             var result = await modal.Result;
 
-            if (result.Cancelled)
+           if (result.Cancelled)
             {
                 return;
             }
 
-            await DataService.Delete(id);
-
+            if (DataService != null)
+            {
+                await DataService.Delete(id);
+            }
             // Reload the page
             NavigationManager.NavigateTo("list", true);
         }
